Hopefully fix obscure bug that's hard to duplicate
Occasionally, the game doesn't end if lives reach zero. I've gotten down to -100 lives before the yeti ate me.
I think the issue is that we were just checking if lives are equal to 0, but if the player hits multiple
obstacles at once, it will just skip past zero and decrement to -10 or whatever the number is.
Changing the logic to check if lives are less than or equal to zero might fix that, we'll see if the bug comes up again